Remove more EF logic out of StandardDataSource`1
